(#7) Add hoook for executing pester tests
This commit adds a hook package to this repository which executes any
pester tests that have been included in a tests folder inside the package.

Package parameters are provided to allow skipping of execution of the pester
tests to allow for other hooks that may be required to run as opposed to using
the global --skip-hooks parameter.